kandu/bootstrap/requirements.txt
locustio
pyzmq 
